(PC-30906)[PRO] feat: Separate collective and individual offer table rows.
