+++ pro/src/components/OfferEducationalActions/OfferEducationalActions.tsx
@@ -27,7 +27,7 @@ import { useNotification } from 'hooks/useNotification'
 import fullHideIcon from 'icons/full-hide.svg'
 import fullNextIcon from 'icons/full-next.svg'
 import strokeCheckIcon from 'icons/stroke-check.svg'
-import { getCollectiveStatusLabel } from 'pages/Offers/Offers/OfferItem/Cells/CollectiveOfferStatusCell/CollectiveOfferStatusCell'
+import { getCollectiveStatusLabel } from 'pages/Offers/Offers/OfferRow/Cells/CollectiveOfferStatusCell/CollectiveOfferStatusCell'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
 import { ButtonVariant, IconPositionEnum } from 'ui-kit/Button/types'
