+++ pro/src/pages/Offers/Offers/OfferRow/Cells/CheckboxCell.tsx
@@ -1,5 +1,3 @@
-import { CollectiveOfferStatus, OfferStatus } from 'apiClient/v1'
-import { isOfferDisabled } from 'core/Offers/utils/isOfferDisabled'
 import { BaseCheckbox } from 'ui-kit/form/shared/BaseCheckbox/BaseCheckbox'
 
 import styles from '../OfferItem.module.scss'
@@ -7,15 +5,13 @@ import styles from '../OfferItem.module.scss'
 interface CheckboxCellProps {
   offerName: string
   isSelected: boolean
-  status: OfferStatus | CollectiveOfferStatus
-  disabled: boolean
+  disabled?: boolean
   selectOffer: () => void
 }
 
 export const CheckboxCell = ({
   offerName,
   isSelected,
-  status,
   disabled,
   selectOffer,
 }: CheckboxCellProps) => {
@@ -26,7 +22,7 @@ export const CheckboxCell = ({
         className="select-offer-checkbox"
         label={offerName}
         exceptionnallyHideLabelDespiteA11y={true}
-        disabled={disabled || isOfferDisabled(status)}
+        disabled={disabled}
         onChange={selectOffer}
       />
     </td>
