+++ pro/src/pages/Offers/Offers/OfferRow/Cells/CollectiveActionsCells.tsx
