+++ pro/src/pages/Offers/Offers/OfferRow/Cells/CollectiveOfferStatusCell/CollectiveOfferStatusCell.tsx
