+++ pro/src/pages/Offers/Offers/OfferRow/Cells/CollectiveOfferStatusCell/__specs__/CollectiveOfferStatusCell.spec.tsx
