+++ pro/src/pages/Offers/Offers/OfferRow/Cells/DeleteDraftCell.tsx
