+++ pro/src/pages/Offers/Offers/OfferRow/Cells/DuplicateOfferCell/DuplicateOfferDialog/DuplicateOfferDialog.module.scss
