+++ pro/src/pages/Offers/Offers/OfferRow/Cells/EditStocksCell.tsx
