+++ pro/src/pages/Offers/Offers/OfferRow/Cells/OfferNameCell/OfferNameCell.tsx
