+++ pro/src/pages/Offers/Offers/OfferRow/Cells/OfferNameCell/__specs__/OfferNameCell.spec.tsx
