+++ pro/src/pages/Offers/Offers/OfferRow/Cells/OfferNameCell/__specs__/utils.offernamecell.spec.tsx
