+++ pro/src/pages/Offers/Offers/OfferRow/Cells/OfferNameCell/utils.ts
