+++ pro/src/pages/Offers/Offers/OfferRow/Cells/OfferRemainingStockCell.tsx
