+++ pro/src/pages/Offers/Offers/OfferRow/Cells/OfferVenueCell.tsx
