+++ pro/src/pages/Offers/Offers/OfferRow/Cells/ThumbCell.tsx
