+++ pro/src/pages/Offers/Offers/OfferRow/Cells/__specs__/CollectiveActionsCells.spec.tsx
