+++ pro/src/pages/Offers/Offers/OfferRow/OfferItem.module.scss
