+++ pro/src/pages/Offers/Offers/OffersTableBody/CollectiveOffersTableBody.tsx
@@ -1,0 +1,39 @@
+import { CollectiveOfferResponseModel } from 'apiClient/v1'
+import { SearchFiltersParams } from 'core/Offers/types'
+import { isSameOffer } from 'pages/Offers/utils/isSameOffer'
+
+import { CollectiveOfferRow } from '../OfferRow/CollectiveOfferRow'
+
+type CollectiveOffersTableBodyProps = {
+  offers: CollectiveOfferResponseModel[]
+  selectOffer: (offer: CollectiveOfferResponseModel) => void
+  selectedOffers: CollectiveOfferResponseModel[]
+  urlSearchFilters: SearchFiltersParams
+}
+
+export const CollectiveOffersTableBody = ({
+  offers,
+  selectOffer,
+  selectedOffers,
+  urlSearchFilters,
+}: CollectiveOffersTableBodyProps) => {
+  return (
+    <tbody className="offers-list">
+      {offers.map((offer) => {
+        const isSelected = selectedOffers.some((selectedOffer) =>
+          isSameOffer(selectedOffer, offer)
+        )
+
+        return (
+          <CollectiveOfferRow
+            isSelected={isSelected}
+            key={`${offer.isShowcase ? 'T-' : ''}${offer.id}`}
+            offer={offer}
+            selectOffer={selectOffer}
+            urlSearchFilters={urlSearchFilters}
+          />
+        )
+      })}
+    </tbody>
+  )
+}
